Rististemate un paio di cose! Guardate i changelog! :)

+++ kernel/intr/irq.c
@@ -82,23 +82,23 @@ void irq_install()
     irq_remap();
 
     // Master
-    idt_set_gate(32, (unsigned)_irq0, 0x08, 0x8E);
-    idt_set_gate(33, (unsigned)_irq1, 0x08, 0x8E);
-    idt_set_gate(34, (unsigned)_irq2, 0x08, 0x8E);
-    idt_set_gate(35, (unsigned)_irq3, 0x08, 0x8E);
-    idt_set_gate(36, (unsigned)_irq4, 0x08, 0x8E);
-    idt_set_gate(37, (unsigned)_irq5, 0x08, 0x8E);
-    idt_set_gate(38, (unsigned)_irq6, 0x08, 0x8E);
-    idt_set_gate(39, (unsigned)_irq7, 0x08, 0x8E);
+    idt_set_gate(32, (size_t)_irq0, 0x08, 0x8E);
+    idt_set_gate(33, (size_t)_irq1, 0x08, 0x8E);
+    idt_set_gate(34, (size_t)_irq2, 0x08, 0x8E);
+    idt_set_gate(35, (size_t)_irq3, 0x08, 0x8E);
+    idt_set_gate(36, (size_t)_irq4, 0x08, 0x8E);
+    idt_set_gate(37, (size_t)_irq5, 0x08, 0x8E);
+    idt_set_gate(38, (size_t)_irq6, 0x08, 0x8E);
+    idt_set_gate(39, (size_t)_irq7, 0x08, 0x8E);
     // Slave
-    idt_set_gate(40, (unsigned)_irq8, 0x08, 0x8E);
-    idt_set_gate(41, (unsigned)_irq9, 0x08, 0x8E);
-    idt_set_gate(42, (unsigned)_irq10, 0x08, 0x8E);
-    idt_set_gate(43, (unsigned)_irq11, 0x08, 0x8E);
-    idt_set_gate(44, (unsigned)_irq12, 0x08, 0x8E);
-    idt_set_gate(45, (unsigned)_irq13, 0x08, 0x8E);
-    idt_set_gate(46, (unsigned)_irq14, 0x08, 0x8E);
-    idt_set_gate(47, (unsigned)_irq15, 0x08, 0x8E);
+    idt_set_gate(40, (size_t)_irq8, 0x08, 0x8E);
+    idt_set_gate(41, (size_t)_irq9, 0x08, 0x8E);
+    idt_set_gate(42, (size_t)_irq10, 0x08, 0x8E);
+    idt_set_gate(43, (size_t)_irq11, 0x08, 0x8E);
+    idt_set_gate(44, (size_t)_irq12, 0x08, 0x8E);
+    idt_set_gate(45, (size_t)_irq13, 0x08, 0x8E);
+    idt_set_gate(46, (size_t)_irq14, 0x08, 0x8E);
+    idt_set_gate(47, (size_t)_irq15, 0x08, 0x8E);
 }
 
 void _irq_handler(struct regs *r)
